arreglando style

+++ src/views/post.js
@@ -1,5 +1,5 @@
-// import { currentUser } from '../firebase/autenticacion.js';
-import { deletePost, editPost } from '../views-controllers/post-control.js';
+import { currentUser } from '../firebase/autenticacion.js';
+import { editLikes, deletePost, editPost } from '../views-controllers/post-control.js';
 // import { editLikes, deletePost, deleteLike, getAllLikes}
 // from '../views-controllers/post-control.js';
 
@@ -47,19 +47,17 @@ export const sharingPost = (data) => {
                           <section class="form-save">
                             <form class="form-save" maxlength="50" required>
                             <textarea class="textarea-post" id="text-post" disabled>${data.postText}</textarea>
-
+                              <section class="heart-commet">
                             <button id="like-${data.id}" class="bottom-heart">
-                            <i class="fa fa-heart heart-full" aria-hidden="true" id="dislike-${data.id}" i>
+                            <i id="count-Like" class="fa fa-heart-o heart-empty" aria-hidden="true">  ${data.likes}</i>
                             </button> 
-                            <button>
-                            <i id="counter-${data.id}" class="fa fa-heart-o heart-empty" aria-hidden="true">  ${data.likes}</i>
-                            </button>
                             <button class="show-comment">
                                 <span id="show-comment">
                                   <i class="fa fa-comment-o show-comment" aria-hidden="true"></i>
                                 </span>
                                 <a class="counter-comment">2</a>
                             </button>
+                            </section>
                             <!-- <span class="margin-left hide">
                                 <i class="fa fa-heart-floppy-o iconSave" aria-hidden="true"></i>
                                 <span></span>
@@ -111,33 +109,43 @@ export const sharingPost = (data) => {
 
   // getAllLikes(data.id, callbackLikes);
 
+  const btnLike = sectionPost.querySelector(`#like-${data.id}`);
   const deletedPost = sectionPost.querySelector('#deletePost');
   const editedPost = sectionPost.querySelector(`#edit-${data.id}`);
   const savePost = sectionPost.querySelector('#savePost');
   const textToEdit = sectionPost.querySelector('#text-post');
 
-  deletedPost.addEventListener('click', () => {
-    deletePost(data.id);
-  });
-
-  editedPost.addEventListener('click', () => {
-    savePost.classList.remove('hide');
+  if (data.idUser !== currentUser().uid) {
+    deletedPost.classList.add('hide');
     editedPost.classList.add('hide');
-    textToEdit.disabled = false;
-    textToEdit.select();
-  });
+  } else {
+    deletedPost.addEventListener('click', () => {
+      deletePost(data.id);
+    });
 
-  savePost.addEventListener('click', () => {
-    editedPost.classList.remove('hide');
-    savePost.classList.add('hide');
-    editPost(data.id, textToEdit.value);
-    textToEdit.disabled = true;
-  });
+    deletedPost.addEventListener('click', () => {
+      deletePost(data.id);
+    });
 
-  // btnLike.addEventListener('click', () => {
-  //   const likeValue = data.likes + 1;
-  //   editLikes(data.id, likeValue);
-  // });
+    editedPost.addEventListener('click', () => {
+      savePost.classList.remove('hide');
+      editedPost.classList.add('hide');
+      textToEdit.disabled = false;
+      textToEdit.select();
+    });
+
+    savePost.addEventListener('click', () => {
+      editedPost.classList.remove('hide');
+      savePost.classList.add('hide');
+      editPost(data.id, textToEdit.value);
+      textToEdit.disabled = true;
+    });
+  }
+
+  btnLike.addEventListener('click', () => {
+    const likeValue = data.likes + 1;
+    editLikes(data.id, likeValue);
+  });
 
   return sectionPost;
 };
